[#85] Move auth pages to public routes and apply new login/forgot/reset design
